Add discussion for each problem
Add a short description of each problem and how I solved it.
This is mostly for myself.
I often find myself wanting to copy a code snippet from one problem, but
I can't remember which day it was from.
Instead of checking each problem page or `grep`ing the repo,
I added this description so I could just search this page for keywords.
(ok it's not the only reason)